feat: move header component
